# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /applicants:
    get:
      operationId: listApplicants
      tags: [applicants]
      summary: Applicant directory
      description: Any active staff member. Newest first.
      responses:
        "200":
          description: Applicants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicantEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /applicants/{id}/tier:
    parameters:
      - $ref: "#/components/parameters/ApplicantId"
    post:
      operationId: setApplicantTier
      tags: [applicants]
      summary: Change an applicant's tier
      description: Requires accounts.tier. A reason is required; the applicant is told.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TierChange"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Stale"
...
  /applicants/{id}/credential-reset:
    parameters:
      - $ref: "#/components/parameters/ApplicantId"
    post:
      operationId: requireCredentialReset
      tags: [applicants]
      summary: Require a new password or two-step setup
      description: Requires accounts.manage. Recorded for the applicant to complete.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialReset"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Stale"
  /applicants/{id}/identity/approve:
    parameters:
      - $ref: "#/components/parameters/ApplicantId"
    post:
      operationId: approveIdentityCheck
      tags: [applicants]
      summary: Approve an identity check
      description: Requires kyc.review.
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Stale"
  /applicants/{id}/identity/reject:
    parameters:
      - $ref: "#/components/parameters/ApplicantId"
    post:
      operationId: rejectIdentityCheck
      tags: [applicants]
      summary: Reject an identity check
      description: Requires kyc.review. The reason is shown to the applicant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reason"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Stale"
  /applicants/{id}/identity/reverify:
    parameters:
      - $ref: "#/components/parameters/ApplicantId"
    post:
      operationId: requestReverification
      tags: [applicants]
      summary: Ask a verified applicant to verify again
      description: Requires kyc.review. The reason is shown to the applicant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reason"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantResult"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Stale"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [activity]
      summary: Your notifications
      description: Newest first (up to 200).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [activity]
      summary: Mark all your notifications read
      description: Marks every one of your notifications as read.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}/read:
    parameters:
      - $ref: "#/components/parameters/NotificationId"
    post:
      operationId: markNotificationRead
      tags: [activity]
      summary: Mark one notification read
      description: Only your own; anyone else's reads as not found.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /staff-feed:
    get:
      operationId: listStaffFeed
      tags: [activity]
      summary: Team activity
      description: Any active staff member. Newest first (up to 200), with your own read state.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffFeedItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /staff-feed/read-all:
    post:
      operationId: markAllStaffFeedRead
      tags: [activity]
      summary: Mark all team activity read (for you)
      description: Marks every activity item as read for you only; other staff keep their own read state.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /staff-feed/{id}/read:
    parameters:
      - $ref: "#/components/parameters/FeedId"
    post:
      operationId: markStaffFeedRead
      tags: [activity]
      summary: Mark one activity item read (for you)
      description: For you only; other staff keep their own read state.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /audit:
    get:
      operationId: getAuditLog
      tags: [activity]
      summary: Audit log
      description: Requires audit.view. Every staff action, newest first, with whether the hash chain is intact.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase Auth access token
  parameters:
    NotificationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^NT-[0-9]{1,15}$"
    FeedId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^FD-[0-9]{1,15}$"
...
    ApplicantId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Invalid:
      description: The change was refused; `fieldErrors` names the fields to fix
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Stale:
      description: The record changed since it was read; reload it and try again
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Missing, invalid, or expired sign-in token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Signed in, but not allowed to do this
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    NotificationItem:
      type: object
      properties:
        id:
          type: string
        at:
          type: string
        title:
          type: string
        body:
          type: string
        href:
          type: string
        read:
          type: boolean
      required: [id, at, title, body, href, read]
    StaffFeedItem:
      type: object
      properties:
        id:
          type: string
        at:
          type: string
        kind:
          type: string
          enum: [application, deposit, withdrawal, card, security, account]
        title:
          type: string
        body:
          type: string
        href:
          type: string
        highlight:
          type: boolean
        read:
          type: boolean
      required: [id, at, kind, title, body, href, highlight, read]
    AuditItem:
      type: object
      properties:
        id:
          type: string
        at:
          type: string
        staffId:
          type: string
        staffName:
          type: string
        role:
          type: string
        action:
          type: string
        target:
          type: string
        applicantId:
          type: string
          nullable: true
        summary:
          type: string
        changes:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              before:
                type: string
              after:
                type: string
            required: [field, before, after]
        riskScore:
          type: integer
          nullable: true
        ip:
          type: string
          nullable: true
      required: [id, at, staffId, staffName, role, action, target, applicantId, summary, changes, riskScore, ip]
    AuditLog:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/AuditItem"
        chain:
          type: object
          description: Whether every entry still matches its hash and links to the one before
          properties:
            intact:
              type: boolean
            checked:
              type: integer
            brokenAt:
              type: string
          required: [intact, checked]
      required: [events, chain]
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
      required: [error]
    Message:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Profile:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        sector:
          type: string
        country:
          type: string
        tier:
          type: integer
          enum: [1, 2, 3]
        identityVerified:
          type: boolean
        joined:
          type: string
          description: ISO date the profile was created
        account:
          $ref: "#/components/schemas/Account"
      required: [name, email, phone, address, sector, country, tier, identityVerified, joined, account]
    Kyc:
      type: object
      properties:
        status:
          type: string
          enum: [Not submitted, Pending, Verified, Rejected]
        documentType:
          type: string
          enum: [Passport, National ID, Driver's licence]
        documentLast4:
          type: string
        nameOnDocument:
          type: string
        submittedAt:
          type: string
        reviewedAt:
          type: string
        reviewedBy:
          type: string
        rejectionReason:
          type: string
      required: [status]
    Account:
      type: object
      description: Staff-managed controls on an applicant account
      properties:
        status:
          type: string
          enum: [Active, Locked]
        lockReason:
          type: string
        lockedAt:
          type: string
        lockedBy:
          type: string
        passwordResetRequired:
          type: boolean
        twoFactorResetRequired:
          type: boolean
        kyc:
          $ref: "#/components/schemas/Kyc"
      required: [status, passwordResetRequired, twoFactorResetRequired, kyc]
    ApplicantEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        profile:
          $ref: "#/components/schemas/Profile"
      required: [id, profile]
    ApplicantResult:
      type: object
      properties:
        applicant:
          $ref: "#/components/schemas/ApplicantEntry"
        message:
          type: string
      required: [applicant, message]
    Reason:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000
      required: [reason]
      additionalProperties: false
    TierChange:
      type: object
      properties:
        tier:
          type: integer
          enum: [1, 2, 3]
        reason:
          type: string
          maxLength: 1000
      required: [tier, reason]
      additionalProperties: false
    CredentialReset:
      type: object
      properties:
        kind:
          type: string
          enum: [password, twoFactor]
      required: [kind]
      additionalProperties: false
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...