# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditItem } from './auditItem';
import type { AuditLogChain } from './auditLogChain';

export interface AuditLog {
  events: AuditItem[];
  /** Whether every entry still matches its hash and links to the one before */
  chain: AuditLogChain;
}